Add "calendar input element" example to cookbook
This one is a bit different, since it runs in the cookbook page directly
in the browser.

Ideally it would be in its own file which would be transcluded both into
the <script> tag and into a code block, but I couldn't get that to work
with the markdown processor, it insisted on interpreting the contents of
the <script> tag as markdown. So instead we add an extra script that
takes the contents of the <script> tag and copies it into the code
block.

See: #240